requirements.txt
flask==2.0.1
werkzeug==2.0.3
torch==2.0.0
torchvision==0.15.1
numpy==1.24.0
Pillow==9.5.0
prometheus-flask-exporter==0.22.4
